extract logic into useCalculateBasePrice & useCalculateTotal; rename dir
